FixCoasreGridCommand() does no load balancing any more. Load balancing should be done by
  ptest $3 <level>;

where <level> is the level to start grid distribution.
e.g. use ptest $3 1, to leave the coarse grid on the master.

or

  LB=<method>;
  ex lb4;

this calls Chaco specifying a lb-method, take LB=3 (RIB),
as default value.
To specify the start level here edit the variable STARTLEVEL
in lb4.scr. This script resides in UG/fe/appl - not the best
place, sorry.

[[Imported from SVN: r2542]]