shell script added:
syntax:

	tstat [-a | -m | -d | -h] <film-name> [<out-name>]

analyses all existing data-files: <film-name>.ug.data.* and writes a gnuplot-file (2D),
to be display with

	gnuplot;
	plot "<out-name>" w l

It shows the time development of the serie of data-files. It can be scaled to show years(-a),
month(-m), days(-d), hours (-h) or without, just the time assumed to be given in seconds.
The out-file name is optional, default is "out.gnu".

[[Imported from SVN: r2666]]