new 'option' for xugv: If xugv is used for one file without options, like 		xugv meta
and there is no meta-file 'meta' but, say

		meta.0000 ... metas.0012

the last 'meta.0012' is opened, if there exists a file '.xugv_tail'
in the home-directory. In that case, '.xugv_tail' is removed. Again,
if a new file 'meta.0013' pops up and '.xugv_tail' exists again, this
meta-file is display next.

This is an easy implementation of the tail-functionality for a meta-file-
film. It can be used, e.g. on speedo. The only thing you need to do is
deleting the file '.xugv_tail' in HOME before plotting a new meta-file,
and touching it after plotting has finished.

[[Imported from SVN: r7158]]