[Buildsystem] Build the libraries in a hierarchical fashion.
This has the effect of not including libparutil.la in liblow?.la.  This is
especially important without --enable-parallel, because in this case
libparutil.la did reference stuff from libparallel.la, which was however not
build.  Now libparutil is included in libparallel and they are both either
build or not build.

[[Imported from SVN: r8367]]