Speed up creation of domain edges slightly by smarter usage of std::set.
Instead of search for an element z in a set A, and inserting it afterwards
only if z hasn't been found, we may insert z directly.  The return value
of the insert() method tells us whether z has actually been added to
the set, or whether it had already been there.

This replace two searches in the set by a single one.

[[Imported from SVN: r8379]]