Change behavior of the insertElement method
InsertElement needs to find the neighbors of a newly inserted element.
Previously, there were two methods: a linear (in time) method which
needs preallocated memory, and a quadratic method with less memory
requirements.  If the grid was small enough, then the linear
algorithm was used, otherwise, it was the quadratic one.
However, the quadratic one takes such a long time as to be effectively
useless.  Hence this patch removes the quadratic algorithm.
If there is not enough space for the linear algorithm, an error
message is produced, which contains instructions on how to increase
the available memory.

[[Imported from SVN: r8429]]