Register the Dune-only type data members with ddd.
DDD needs to know about the structure of the data types it is
supposed to transfer.  We never told it about the things we added
when compiling for Dune.  This never seemed to cause any trouble,
I guess because all Dune-only data is local anyway.
This patch still adds the extra information.  You never know :-)

[[Imported from SVN: r8439]]