Introduce a configure flag --enable-system-heap
If this flag is set UG takes it memory through malloc/free instead of using
the custom heap data structures in low/heaps.c.

Using the system heap has several advantages:
 - there is no artificial upper bound to the amount of memory UG can use
 - valgrind may see more bugs

The system heap may also be faster, but it may also be slower.
No way to know without measuring.

This patch is incomplete: not all memory that UG currently tells the
internal heap to free is freed on the system heap, too.  Hence
running UG with --enable-system-heap now will result in various
memory leaks.

Since this is an experimental feature it is switched off by default

[[Imported from SVN: r8472]]