Remove hack in header installation
This Makefile contained a hack from the times when the AutoTools
build system for UG was still incomplete: for header installation,
all header files that 'find' would find in the source tree were
simply copied to the installation directory.  But now the AutoTools
system is in good shape and the headers can be installed in the
standard way.  Therefore the old hack can be removed.

[[Imported from SVN: r8491]]