Remove the artificial 3d-array of processors set up in the ppif layer
There is no guarantee that it matches the topology of the actual machine
in any way.  Also, it does not appear to be used anywhere in the ug core
at all.  Finally, it contains some small memory leaks.  Admittedly, these
leaks would be easy to fix, but since the code is not used I may as well
rip it out altogether.

[[Imported from SVN: r8530]]