[bugfix] Dispose doubly created SideVectors in RefineElementGreen
When UG creates a new element, and SideVectors are switched on,
then a new SideVector is created for each face of the new element.
Since most faces are shared by two elements, though, the consequence
is that there are two SideVectors for each interior face.  To obtain
a single SideVector per face (as it should be), the UG refinement
code calls the method DisposeDoubledSideVector for these duplicated
SideVectors.  This appears to be easier than only creating one
SideVector for each face to begin with.

However, so far the duplicate SideVectors were not disposed in a
green, non-rule based refinement of an element.  This lead to
memory leaks which could be observed when running the Dune UGGrid
test with --enable-system-heap and valgrind.  This patch adds
the missing calls to DisposeDoubledSideVector, and fixes those
leaks.

Incidentally, this also appears to be the proper fix for the issue
described in Dune FlySpray 170,

https://dune-project.org/flyspray/index.php?do=details&task_id=170

which had previously been hacked around in dune-grid.