Remove code that redefines standard types on specific architectures
This is a left-over from the old build system.  For some explicitly given
architectures, UINT is not unsigned int, for example.  The special-casing
stopped working when we moved to the AutoTools long time ago, but hardly
anybody noticed because amd64 worked anyway.  Now this old code gives
us trouble building dune-grid on Debian ppc64el.  Since I don't even
know why these types need to be changed, and what the requirements on
the types are, I just kick out all the special cases.  If something
goes wrong we can then fix it in a portable way.